Update Graham Scan
